fix: camelcaseKeys does not support CommonJS
move the pkgroll to only generate ESM files, not CommonJS
this fixes the error "TypeError: camelcaseKeys is not a function"

fix: circular references
